inventory-backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-versatileimagefield==2.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0